requirements.txt
black==22.8.0
click==8.1.3
grpcio==1.49.1
grpcio-tools==1.49.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
protobuf==4.21.6
six==1.16.0
tomli==2.0.1
